JAMES-2586 Move some class related to ACLUpdated event from apache-james-mailbox-cassandra -> apache-james-mailbox-event-json
- In order to make it more generic, that can re-usable
