+++ mailbox/cassandra/src/main/java/org/apache/james/mailbox/cassandra/mail/CassandraACLMapper.java
@@ -25,6 +25,8 @@ import jakarta.inject.Inject;
 
 import org.apache.commons.lang3.NotImplementedException;
 import org.apache.james.backends.cassandra.init.configuration.CassandraConfiguration;
+import org.apache.james.event.MailboxAggregateId;
+import org.apache.james.event.acl.ACLUpdated;
 import org.apache.james.eventsourcing.Command;
 import org.apache.james.eventsourcing.CommandHandler;
 import org.apache.james.eventsourcing.EventSourcingSystem;
@@ -32,10 +34,8 @@ import org.apache.james.eventsourcing.Subscriber;
 import org.apache.james.eventsourcing.eventstore.EventStore;
 import org.apache.james.mailbox.acl.ACLDiff;
 import org.apache.james.mailbox.cassandra.ids.CassandraId;
-import org.apache.james.mailbox.cassandra.mail.eventsourcing.acl.ACLUpdated;
 import org.apache.james.mailbox.cassandra.mail.eventsourcing.acl.AclV2DAOSubscriber;
 import org.apache.james.mailbox.cassandra.mail.eventsourcing.acl.DeleteMailboxCommand;
-import org.apache.james.mailbox.cassandra.mail.eventsourcing.acl.MailboxAggregateId;
 import org.apache.james.mailbox.cassandra.mail.eventsourcing.acl.SetACLCommand;
 import org.apache.james.mailbox.cassandra.mail.eventsourcing.acl.UpdateACLCommand;
 import org.apache.james.mailbox.cassandra.mail.eventsourcing.acl.UserRightsDAOSubscriber;
