+++ mailbox/cassandra/src/main/java/org/apache/james/mailbox/cassandra/mail/eventsourcing/acl/UserRightsDAOSubscriber.java
@@ -19,9 +19,11 @@
 
 package org.apache.james.mailbox.cassandra.mail.eventsourcing.acl;
 
+import org.apache.james.event.acl.ACLUpdated;
 import org.apache.james.eventsourcing.Event;
 import org.apache.james.eventsourcing.EventWithState;
 import org.apache.james.eventsourcing.ReactiveSubscriber;
+import org.apache.james.mailbox.cassandra.ids.CassandraId;
 import org.apache.james.mailbox.cassandra.mail.CassandraUserMailboxRightsDAO;
 import org.reactivestreams.Publisher;
 
@@ -39,7 +41,7 @@ public class UserRightsDAOSubscriber implements ReactiveSubscriber {
         Event event = eventWithState.event();
         if (event instanceof ACLUpdated) {
             ACLUpdated aclUpdated = (ACLUpdated) event;
-            return userRightsDAO.update(aclUpdated.mailboxId(), aclUpdated.getAclDiff());
+            return userRightsDAO.update((CassandraId) aclUpdated.mailboxId(), aclUpdated.getAclDiff());
         }
         return Mono.empty();
     }
